Fix: ProfileMBTIViewModel
