+++ NaMuApp/Feat/Profile/ProfileMBTI/ProfileMBTIViewModel.swift
@@ -6,9 +6,27 @@
 //
 
 import Foundation
+import UIKit
 
-final class ProfileMBTIViewModel {
-    let mbti = ProfileViewModel.MbtiType.allCases
+final class ProfileMBTIViewModel: ViewModelType {
+    
+    enum MbtiType: CaseIterable {
+        case IE
+        case NS
+        case FT
+        case PJ
+    }
+    
+    let mbti = ProfileMBTIViewModel.MbtiType.allCases
+    var tapped: (()->Void)?
+    
+    struct Input {
+        
+    }
+    
+    struct Output {
+        
+    }
     
     init() {
         print(#function, self)
@@ -17,4 +35,30 @@ final class ProfileMBTIViewModel {
     deinit {
         print(#function, self)
     }
+    
+    func checkMBTI(_ bools: [Bool?]) -> String? {
+        var returnString = ""
+        for (index, bool) in bools.enumerated() {
+            guard let valid = bool else { return nil }
+            switch MbtiType.allCases[index] {
+            case .IE:
+                returnString += (valid ? "E" : "I")
+            case .NS:
+                returnString += (valid ? "S" : "N")
+            case .FT:
+                returnString += (valid ? "T" : "F")
+            case .PJ:
+                returnString += (valid ? "J" : "P")
+            }
+        }
+        return returnString
+    }
+}
+
+extension ProfileMBTIViewModel {
+    
+    func transform(input: Input) -> Output {
+        return Output()
+    }
+    
 }
